+++ src/results.jl
@@ -1,12 +1,13 @@
 import JSON
 
 """Result of a Carlo Monte Carlo calculation containing the mean, statistical error and autocorrelation time."""
-mutable struct ResultObservable{T<:Number,R<:Real,N,M}
+mutable struct ResultObservable{T<:Number,R<:Real,N,M,C<:Union{<:AbstractArray,Nothing}}
     internal_bin_length::Int64
     rebin_length::Int64
 
     mean::Array{T,N}
     error::Array{R,N}
+    covariance::C
     autocorrelation_time::Array{R,N}
 
     rebin_means::Array{T,M}
@@ -17,6 +18,7 @@ rebin_count(obs::ResultObservable) = Int64(size(obs.rebin_means)[end])
 JSON.lower(obs::ResultObservable) = Dict(
     "mean" => obs.mean,
     "error" => obs.error,
+    "covariance" => obs.covariance,
     "autocorr_time" => maximum(obs.autocorrelation_time),
     "rebin_len" => obs.rebin_length,
     "rebin_count" => rebin_count(obs),
@@ -39,7 +41,9 @@ function write_results(
                 "parameters" => parameters,
                 "results" => observables,
                 "version" => to_dict(version),
-            ); pretty = 1, allownan = true
+            );
+            pretty = 1,
+            allownan = true,
         )
     end
     return nothing